Improve parser method logic by simplifying
The logic for the parser methods has been simplified, to allow
stricter following of the format rules, and to prevent bugs.

The new logic processes the input string from left-to-right, and
works forward on each step, rather than trying to evaluate the
whole string at once. This allows for the identification of data
elements to be much less complex.