Modify ParserInterface.parse $flags argument
Keep the interfaces consistent in design for the Parser and
Formatter logic to avoid confusion.

Change the type requirement for the $flags argument of the parse
method to allow for null values. This will allow for a sane
default value to be used, instead of an empty array.

An empty array could be provided if the user doesn't provide any
flags, but this would override the "default" flags set earlier.

The new null default value removes the confusion between the two
options. NULL is used when there are no NEW flags requested, i.e.
use the already set flags. An empty array is used to remove all of
the previously set flags.