Add exception to ParserInterface.parse & fix JsonParser
Add a new exception: InvalidArgumentException to the parse()
method of the ParserInterface, which gets thrown when the parser
is unable to parse the provided input argument.

This commit also fixes the interface rule breaking of the
JsonParser class. Previously false would be returned on failure to
parse, which is not an allowed return type in the ParserInterface.