Change TaskItem to use DateTime objects for dates
Redesign the logic of the creation date and completion date
methods to make use of DateTime objects instead of "YYYY-MM-DD"
formatted date strings.

The "raw" class property is stored as an ISO8601 string and is
converted to and from a DateTime object as needed. This design
decision was made to allow for ORMs to be able to work with the
DateTime values better, especially as DateTime has no __toString()
method.

This change will allow for much more complex and powerful date
operations on the TaskItem object. Previously, the design of the
class was too strongly "coupled" to the design of a todo.txt
style of item. The new logic allows for new time values to be
provided instead of just a date.